First release version

+++ TemplateExpress/CreateFiles/CreatePackageFile.cs
@@ -12,23 +12,24 @@ static class CreatePackageJson
         string packageContent = AllFilesText.Package(Options.OnlyJs);
 
         string directory = Directory.GetCurrentDirectory();
-        
+
         //criar novo projeto
-        directory += $@"\{Options.OutputName}";
+        if (!Options.CreateOnCurrentDir)
+            directory += $@"\{Options.OutputName}";
 
         //ver se já existe
-        if (Directory.Exists(directory) && !dev && !directory.EndsWith("TemplateExpress") && !directory.EndsWith("TemplateExpress\\"))
+        if (Directory.Exists(directory) && !dev && !directory.EndsWith("TemplateExpress") && !directory.EndsWith("TemplateExpress\\") && !Options.CreateOnCurrentDir)
         {
-            if(directory.Contains("Template"))
-            Console.WriteLine($"[ERROR] The folder \"{Options.OutputName}\" already exists in the current directory");
+            if (directory.Contains("Template"))
+                Console.WriteLine($"[ERROR] The folder \"{Options.OutputName}\" already exists in the current directory");
             bool overwrite = Input.YesOrNo("Overwrite [y/n]: ");
-            if(!overwrite) throw new Exception("[ERROR] Unable to create files");
+            if (!overwrite) throw new Exception("[ERROR] Unable to create files");
 
             Directory.Delete(directory, true);
         }
 
-
-        Directory.CreateDirectory(directory);
+        if (!Options.CreateOnCurrentDir)
+            Directory.CreateDirectory(directory);
 
 
         Logs.CurrentStep("Starting project");
@@ -37,7 +38,7 @@ static class CreatePackageJson
         {
             File.WriteAllText(Path.Combine(directory, "package.json"), packageContent);
         }
-        
+
 
         File.WriteAllText(Path.Combine(directory, "package.json"), packageContent);
